feat: Track video views in Firestore
This commit introduces functionality to track video views in Firestore.

When a user scrolls to a new video in the `ReelsVideoScreen`, the `_trackVideoView` function is called. This function attempts to add the current `userId` to the `views` array field of the corresponding video document in the `videos` collection.

`FieldValue.arrayUnion` is used to ensure that a user's ID is only added to the `views` array if it's not already present, preventing duplicate entries. `SetOptions(merge: true)` is used to update the `views` field without overwriting other data in the video document.

This change allows for tracking unique views per video.