Refactor: Display video tags as individual clickable elements
This commit refactors the display of video tags in the `SingleVideoView`.

Previously, all tags were displayed as a single string, with each tag prefixed by '#'. Clicking anywhere on this string would navigate to a search view for all tags.

Now, each tag is rendered as an individual `InkWell` widget wrapped in a `Wrap` layout. This allows each tag to be clicked separately, triggering a search for that specific tag. The `_isTagExpanded` logic has been removed as the `Wrap` widget handles overflowing tags.