Refactor: Navigate back from SignUpView instead of to SignInView
This commit changes the navigation behavior of the "Sign In" link on the `SignUpView`.

Previously, tapping the "Sign In" `TextSpan` would navigate the user to the `SignInView` using `Get.toNamed(AppRoutes.signIn)`.

With this change:
- The navigation is updated to use `Get.back()`.

This assumes that the `SignUpView` is typically pushed onto the navigation stack after the `SignInView`, so `Get.back()` will return the user to the previous `SignInView`. This simplifies the navigation flow.